requirements.txt
alabaster==0.7.12
asgiref==3.5.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-bootstrap-modal-forms==2.2.0
django-crispy-forms==1.14.0
django-filter==1.1.0
django-sphinx-view==22.1a6
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-filters==0.11.1
docutils==0.17.1
idna==3.3
imagesize==1.3.0
Jinja2==3.0.3
MarkupSafe==2.1.0
packaging==21.3
Pygments==2.11.2
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
snowballstemmer==2.2.0
Sphinx==4.4.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-packages==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
urllib3==1.26.8
